+++ src/platform/roblox/RobloxStudioPlugin.cpp
@@ -131,7 +131,7 @@ bool RobloxPlatform::hydrateSourcemapWithPluginInfo()
     if (didUpdateSourcemap)
     {
         // Update the sourcemap file if needed
-        auto config = workspaceFolder->client->getConfiguration(workspaceFolder->rootUri);
+        auto config = workspaceFolder->getConfiguration();
         if (config.sourcemap.enabled && config.sourcemap.autogenerate)
         {
             auto sourcemapPath = workspaceFolder->rootUri.resolvePath(config.sourcemap.sourcemapFile);